chapter8/building_custom_skills_for_azure_ai/requirements.txt
azure-functions
azure-core==1.24.0
azure-ai-formrecognizer==3.2.0
